refactor(adapters): scaffold sections/ module and move header + components renderers
Create the sections/ subdirectory under markdown_formatter/ to begin splitting
the 949-line section.rs into focused single-responsibility modules.

- Add sections/section_renderer.rs: SectionRenderer trait scaffold (used in #428, #429)
- Add sections/header.rs: render_header logic moved here as sections::header::render
- Add sections/components.rs: render_components logic moved here as sections::components::render
- Add sections/mod.rs: re-exports the three submodules
- Update mod.rs: add `mod sections;` and update call sites accordingly
- Remove render_header and render_components from section.rs (4 remaining functions intact)

All 488 tests pass. cargo clippy -- -D warnings clean.

Closes #427
