+++ src/adapters/outbound/formatters/markdown_formatter/sections/mod.rs
@@ -1,0 +1,3 @@
+pub(super) mod components;
+pub(super) mod header;
+pub(super) mod section_renderer;
